+++ crates/zng-task/src/channel/ipc_file.rs
@@ -1,0 +1,193 @@
+use std::{mem, sync::atomic::AtomicUsize};
+
+use serde::{Deserialize, Serialize};
+
+use crate::channel::is_ipc_serialization;
+
+/// File handle that can be transferred to another process.
+///
+/// # File
+///
+/// This type can be converted  from and to [`std::fs::File`]. This type does not
+/// implement IO traits, it will only close the handle on drop if it was not sent or converted.
+///
+/// # Serialization
+///
+/// This type implements serialization only for compatibility with IPC channel, attempting to
+/// serialize or deserialize it outside of [`with_ipc_serialization`] context will panic.
+///
+/// [`with_ipc_serialization`]: crate::channel::with_ipc_serialization
+pub struct IpcFile {
+    handle: AtomicUsize,
+}
+impl From<std::fs::File> for IpcFile {
+    fn from(file: std::fs::File) -> Self {
+        #[cfg(not(any(windows, unix)))]
+        compile_error!("ipc_file mod should not be compiled in this os");
+
+        #[cfg(windows)]
+        let handle = std::os::windows::io::IntoRawHandle::into_raw_handle(file) as usize;
+        #[cfg(unix)]
+        let handle = std::os::fd::IntoRawFd::into_raw_fd(file) as usize;
+        Self {
+            handle: AtomicUsize::new(handle),
+        }
+    }
+}
+impl From<IpcFile> for std::fs::File {
+    fn from(mut f: IpcFile) -> Self {
+        let handle = mem::take(f.handle.get_mut());
+        assert!(handle != 0, "cannot access file, already moved");
+        // SAFETY: handle was not moved (not zero) and was converted from File
+        unsafe { into_file(handle) }
+    }
+}
+impl Drop for IpcFile {
+    fn drop(&mut self) {
+        let handle = mem::take(self.handle.get_mut());
+        if handle != 0 {
+            // SAFETY: handle was not moved (not zero) and was converted from File
+            drop(unsafe { into_file(handle) });
+        }
+    }
+}
+unsafe fn into_file(handle: usize) -> std::fs::File {
+    #[cfg(windows)]
+    unsafe {
+        std::os::windows::io::FromRawHandle::from_raw_handle(handle as _)
+    }
+    #[cfg(unix)]
+    unsafe {
+        std::os::fd::FromRawHandleFd::from_raw_fd(handle as _)
+    }
+}
+impl Serialize for IpcFile {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        assert!(is_ipc_serialization(), "IpcFile can only be serialized inside IPC channels");
+        let handle = self.handle.swap(0, std::sync::atomic::Ordering::Relaxed);
+        assert!(handle != 0, "IpcFile already moved");
+        // SAFETY: handle was not moved (not zero) and was converted from File
+        let handle = unsafe { into_file(handle) };
+
+        #[cfg(windows)]
+        {
+            // -> Sends a channel sender to receive the target process id and a sender to continue the protocol
+            // <- Receives the target process id, DuplicateHandle
+            // -> Sends the new handle and a confirmation sender
+            // <- Receives confirmation, drops this handle
+
+            // ->
+            let (s, mut r) =
+                super::ipc_unbounded::<(u32, super::IpcSender<(usize, super::IpcSender<bool>)>)>().map_err(serde::ser::Error::custom)?;
+            let ok = Serialize::serialize(&s, serializer)?;
+
+            // <-
+            blocking::unblock(move || {
+                let _hold = &handle;
+                if let Ok((process_id, mut shared_sender)) = r.recv_blocking() {
+                    use std::os::windows::io::AsRawHandle as _;
+                    if let Some(handle) = duplicate_handle(process_id, handle.as_raw_handle() as usize) {
+                        // ->
+                        if let Ok((s, mut r)) = super::ipc_unbounded()
+                            && shared_sender.send_blocking((handle, s)).is_ok()
+                        {
+                            // <-
+                            let _ = r.recv_blocking();
+                        }
+                    }
+                    // else sender will disconnect
+                }
+            })
+            .detach();
+            Ok(ok)
+        }
+        #[cfg(unix)]
+        {
+            todo!("https://github.com/standard-ai/sendfd/blob/master/src/lib.rs")
+        }
+
+        #[cfg(not(any(windows, unix)))]
+        {
+            panic!("IpcFile not implemented for {}", std::env::consts::OS);
+        }
+    }
+}
+impl<'de> Deserialize<'de> for IpcFile {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        assert!(is_ipc_serialization(), "IpcFile can only be deserialized inside IPC channels");
+
+        #[cfg(windows)]
+        {
+            type Confirm = bool;
+            type Handle = (usize, super::IpcSender<Confirm>);
+            type Process = (u32, super::IpcSender<Handle>);
+
+            let mut process_id_sender = <super::IpcSender<Process> as Deserialize<'de>>::deserialize(deserializer)?;
+            let (s, mut handle_receiver) = super::ipc_unbounded::<Handle>().map_err(serde::de::Error::custom)?;
+
+            process_id_sender
+                .send_blocking((std::process::id(), s))
+                .map_err(serde::de::Error::custom)?;
+
+            let (handle, mut confirm_sender) = handle_receiver.recv_blocking().map_err(serde::de::Error::custom)?;
+
+            use std::os::windows::io::FromRawHandle as _;
+            // SAFETY: this handle is the output of DuplicateHandle for the current process
+            let handle = unsafe { std::fs::File::from_raw_handle(handle as _) };
+
+            let _ = confirm_sender.send_blocking(true);
+
+            Ok(handle.into())
+        }
+
+        #[cfg(unix)]
+        {
+            todo!()
+        }
+
+        #[cfg(not(any(windows, unix)))]
+        {
+            panic!("IpcFile not implemented for {}", std::env::consts::OS);
+        }
+    }
+}
+
+#[cfg(windows)]
+fn duplicate_handle(process_id: u32, handle: usize) -> Option<usize> {
+    unsafe {
+        use windows_sys::Win32::Foundation::{DUPLICATE_SAME_ACCESS, DuplicateHandle, HANDLE};
+        use windows_sys::Win32::System::Threading::{GetCurrentProcess, OpenProcess, PROCESS_DUP_HANDLE};
+
+        let target_process = OpenProcess(PROCESS_DUP_HANDLE, 0, process_id);
+        if target_process != 0 {
+            let mut target_handle: HANDLE = 0;
+            let success = DuplicateHandle(
+                GetCurrentProcess(),
+                handle as HANDLE,
+                target_process,
+                &mut target_handle,
+                0,
+                0,
+                DUPLICATE_SAME_ACCESS,
+            );
+
+            windows_sys::Win32::Foundation::CloseHandle(target_process);
+
+            if success != 0 {
+                Some(target_handle as usize)
+            } else {
+                tracing::error!("failed to duplicate IpcFile handle");
+                None
+            }
+        } else {
+            tracing::error!("failed to connect to target process for IpcFile handle duplication");
+            None
+        }
+    }
+}